requirements.txt
streamlit
flask
google-cloud-bigquery
pandas
db-dtypes
requests
elasticsearch==8.13.2
pycountry